services/raw_data_processor/requirements.txt
boto3
python-dotenv
pyarrow
